added skip if key pressed instruction and changed keymap key type

+++ model/keyboard.go
@@ -4,7 +4,7 @@ import (
 	"github.com/hajimehoshi/ebiten"
 )
 
-var KeyMap = map[int]ebiten.Key{
+var KeyMap = map[uint16]ebiten.Key{
 	1: ebiten.Key1, 2: ebiten.Key2, 3: ebiten.Key3, 12: ebiten.Key4,
 	4: ebiten.KeyQ, 5: ebiten.KeyW, 6: ebiten.KeyE, 13: ebiten.KeyR,
 	7: ebiten.KeyA, 8: ebiten.KeyS, 9: ebiten.KeyD, 14: ebiten.KeyF,
